backend/services/ocr_service.py
import pytesseract
from PIL import Image
import os

pytesseract.pytesseract.tesseract_cmd = r"C:\Program Files\Tesseract-OCR\tesseract.exe"

def extract_text_from_image(image_path: str) -> str:
    if not os.path.exists(image_path):
        raise FileNotFoundError("Image not found")

    image = Image.open(image_path)
    text = pytesseract.image_to_string(image)

    return text.strip()
