backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
openai
whisper
numpy
